refactor: move tree element types to util/element, extractors to modules/extract
Consolidate all tree/path/code element types and intrinsics into
util/element.ts (browser-safe, pure types) and move extractors to a
top-level modules/extract/ module (build-time only, imports typescript).
Removes modules/tree/ entirely.
