--- modules/tree/element/index.ts
@@ -1,5 +1,0 @@
-export * from "./CodeElement.js";
-export * from "./intrinsics.js";
-export * from "./PathElement.js";
-export * from "./path.js";
-export * from "./TreeElement.js";
