--- modules/tree/index.ts
@@ -1,2 +1,0 @@
-export * from "./element/index.js";
-export * from "./extractor/index.js";
